move the new common main stuff to `cli`
