+++ cmd/klotho/main.go
@@ -1,13 +1,12 @@
 package main
 
 import (
-	"github.com/klothoplatform/klotho/cmd/klothocommon"
 	"github.com/klothoplatform/klotho/pkg/cli"
 	"github.com/klothoplatform/klotho/pkg/updater"
 )
 
 func main() {
-	km := klothocommon.KlothoMain{
+	km := cli.KlothoMain{
 		UpdateStream: updater.DefaultStream,
 		Version:      Version,
 		PluginSetup: func(psb *cli.PluginSetBuilder) error {
